tehtävää 4.22, en ole varma toimiiko vielä oikein

+++ osa 4/blogilista/controllers/blogs.js
@@ -8,6 +8,7 @@ blogsRouter.get('/', async (request, response) => {
   response.json(blogs)
 })
 
+/*
 const getTokenFrom = (request) => {
   const authorization = request.get('authorization')
   if (authorization && authorization.startsWith('bearer ')) {
@@ -15,14 +16,16 @@ const getTokenFrom = (request) => {
   }
   return null
 }
+*/
 
 blogsRouter.post('/', async (request, response, next) => {
   const body = request.body
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+  /*const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token invalid' })
-  }
-  const user = await User.findById(decodedToken.id)
+  }*/
+  const user = request.user
 
   const blog = new Blog({
     author: body.author,
@@ -43,8 +46,16 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  //const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  const user = request.user
+  console.log('user: ', user)
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(404).json({ error: 'Not valid user for delete operation' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
